% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezverse.R
\docType{package}
\name{ezverse-package}
\alias{ezverse}
\alias{ezverse-package}
\title{ezverse: Easily install and load the ezverse}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The ezverse is intended to make various aspects of data science, automated reporting, and data exploration much easier.
}
\author{
\strong{Maintainer}: Eric Lamphere \email{ericjlamphere@gmail.com}

}
\keyword{internal}
