% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ezverse_packages}
\alias{ezverse_packages}
\title{List all packages in the ezverse}
\usage{
ezverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include ezverse in the list?}
}
\description{
List all packages in the ezverse
}
\examples{
ezverse_packages()
}
