% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ezverse_update}
\alias{ezverse_update}
\title{Update ezverse packages}
\usage{
ezverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
ezverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
This will check to see if all ezverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
ezverse_update()
}
}
